Code for Arduino Board
That code could be used with any Arduino or similar board, but this is designed for a Arduino Mega 2560 because the code use the Serial Port to communicate whit the NodeMCu board. From the NodeMCU, the arduino board receives the signal to control the stepper motor to move to the position where the light sensor detects the required light intensity. This signal is received using some digital port as a multiplexor. Other signal is sent trough the serial port with the Temperature and Humiduty data to NodeMCU.